Fix ubuntu exec_test
Test passes on Fedora because the registry server is one of the
defaults.  However it is not typically configured on Ubuntu hosts, and
therefor this test can fail.  While specifying the FQIN in the
dockerfile text is not an ideal solution, it cannot negatively
affect other tests which utilize `podmanTest.BuildImage`.

Signed-off-by: Chris Evich <cevich@redhat.com>