Get safe subnets for integration testing
Currently if someone needs to add a network test with a
specific subnet they have to choose one manually and have
to make sure that this network is not already used. This
is error prone.

Even if the CI passes this can still lead to flakes since
the tests are run parallel and networks could intercept
each other. The proper solution for this is to make sure
that each test uses a globally unique subnet.

This commit implements a GetSafeIPv4Subnet function which
should be used to get a custom subnet for your tests.
The function ensures that the generated subnet is globally
unique across the ginkgo test nodes.

In order for this to work each test which uses a custom
subnet has to use this function. It's up to the maintainers
to make sure new test will use this function.

Signed-off-by: Paul Holzinger <paul.holzinger@web.de>