Fix pull method selection
When using 'podman build --pull=true', the image was not pulled
if the image being pulled was present locally, but a newer version
was in the repository.  It was only pulled if there was no image
in local storage.

In addition, the pull-never option was ignored.  The line
`if flags.Pull{` at line 244 was always returning true
negating the default pullPolicy of PullNever.

Reworked the algorthim for the selection process.  Now
PullIfNewer is set to the default, and then we set the
pullPolicy appropriately based on the other flags
passed in to this routine.

As an FYI, logic run in the calling functions ensures
that we have only one pull flag in the command.

Addresses: #8024

Signed-off-by: TomSweeneyRedHat <tsweeney@redhat.com>