Fix podman search tests to use common registries
An outage caused `registry.redhat.io` to start throwing the error
message indicates a login is now required.  Simplify the hard-coding of
registries for these tests by limiting them to registries we already
rely on extensively elsewhere.  This will make future failures more
obvious, and limit the number of ways in which tests can flake.

Signed-off-by: Chris Evich <cevich@redhat.com>