pull: account for platform
When pulling an image, account for the specified OS, architecture and
variant when looking up local images.  While a local image may be found
based on the specified name, the platform may be different from what the
user desires.  In that case, do not use the local image but continue
pulling.

Also remove pull-policy logic from the client.  That'll reduce one
roundtrip for the remote client and reduces code scattering.  The
backend should be the single source of truth for pull-policy handling.

Fixes: #8001
Signed-off-by: Valentin Rothberg <rothberg@redhat.com>