Reenable 'exec' test on remote
It was 'skip'ped due to frequent flakes (#7241). I just tried
running the 7241 reproducer on my laptop for one hour, and
saw no failures, so let's reenable this in CI and see if it
comes back.

I really hate problems that "go away" on their own without
being explicitly acknowledged and fixed.

Signed-off-by: Ed Santiago <santiago@redhat.com>