[CI:DOCS] Fix Commands table
The markdown table that lists all of the available commands
on the toplevel Podman man page does not resolve correctly
when converted to html and moved to docs.podman.io.

So as html is valid in a markdown document, I've converted
the table to use html instead in hope that it will translate better.

Signed-off-by: TomSweeneyRedHat <tsweeney@redhat.com>